Add manage.py
This script is useful for various tasks relating to the management
of this application.

For example, it allows us to use the `diffsettings` command to review
the current configuration about the Django application.

It also allows us to call into management commands provided by
Juntagrico, which are for example `create_member_for_superusers`,
`generate_depot_list`, `generate_testdata` and `generate_testdata_advanced`,
which sound like they could be useful.