README: Updates for current setup
The README still assumed this repository contained only the Dockerfile
and the docker-compose setup was in a separate repository.

These files are both colocated in this repository by now, so update the
documentation to reflect this.

Also, add a known issues section.

Also, update the instructions for building the image locally, explaining
how it can then be used with the docker-compose setup and how to get
back to "upstream".