start.py: Don't run makemigrations
Creating migrations at startup will lead to migrations being created in
the container that are applied to the database and will be gone when the
container recreated, as they are not part of the container image.

As migrations affect the persistence in the database, they should be
versioned and provided as part of a git tree or a release and not be
created and applied on the fly without review.