WIP: Development docker image
This is a first attempt at a development docker image.

It bind-mounts the juntagrico source code from the host,
allowing developers to test changes immediately without
rebuilding the container.

For this to work, the development stage of the Dockerfile
needs to be built using:

    docker build . --target development -t rotebeete/juntagrico

And a juntagrico repository needs to be checked out next
to the juntagrico-docker tree.